refactor(all): more clean up
